Step 1: Function parameters and return values changed to int.

+++ RubiChess/search.cpp
@@ -2,7 +2,7 @@
 #include "RubiChess.h"
 
 
-short getQuiescence(engine *en, short alpha, short beta, int depth, bool force)
+int getQuiescence(engine *en, int alpha, int beta, int depth, bool force)
 {
     short score;
     bool isLegal;
@@ -86,9 +86,9 @@ short getQuiescence(engine *en, short alpha, short beta, int depth, bool force)
 }
 
 
-short alphabeta(engine *en, short alpha, short beta, int depth, bool nullmoveallowed)
+int alphabeta(engine *en, int alpha, int beta, int depth, bool nullmoveallowed)
 {
-    short score;
+    int score;
     short bestscore = SHRT_MIN + 1;
     chessmove best;
     int eval_type = HASHALPHA;
